pyproject.toml
[project]
name = "ipbes_notifier"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "AntoninRousset", email = "antonin@rousset.com" }]
dependencies = [
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.3",
    "python-telegram-bot[job-queue]>=21.1.1",
    "sqlmodel>=0.0.19",
    "typer>=0.12.3",
    "beautifulsoup4>=4.12.3",
]
requires-python = "==3.11.*"
readme = "README.md"
license = { text = "MIT" }

[tool.pdm.build]
includes = ["src"]
package-dir = "src"

[tool.pdm.dev-dependencies]
dev = ["ruff>=0.4.3", "mypy>=1.10.0"]
